@@ -7,8 +7,8 @@
 @code
 {
     [Parameter]
-    private RenderFragment ChildContent { get; set; }
+    public RenderFragment ChildContent { get; set; }
 
     [Parameter]
-    private bool IsGreyedOut { get; set; }
+    public bool IsGreyedOut { get; set; }
 }
