@@ -50,7 +50,7 @@
             </select>
         </div>
         <div class="ml-auto">
-            <button @onclick="@(() => OnSearch.InvokeAsync(criteria))" type="button" class="btn btn-danger px-5">
+            <button @onclick="() => OnSearch.InvokeAsync(criteria)" type="button" class="btn btn-danger px-5">
                 Search ➝
             </button>
         </div>
@@ -61,7 +61,7 @@
 @code
 {
     [Parameter]
-    private EventCallback<SearchCriteria> OnSearch { get; set; }
+    public EventCallback<SearchCriteria> OnSearch { get; set; }
 
     private SearchCriteria criteria = new SearchCriteria("LHR", "SEA");
 }
