@@ -26,8 +26,8 @@
 @code
 {
     [Parameter]
-    private string Symbol { get; set; }
+    public string Symbol { get; set; }
 
     [Parameter]
-    private FlightSegment Flight { get; set; }
+    public FlightSegment Flight { get; set; }
 }
