@@ -18,5 +18,5 @@
 @code
 {
     [Parameter]
-    private FlightSegment Flight { get; set; }
+    public FlightSegment Flight { get; set; }
 }
