@@ -18,7 +18,7 @@
 
 @code
 {
-    protected override void OnInit()
+    protected override void OnInitialized()
     {
         flightSearchState.OnChange += StateHasChanged;
     }
